+++ extensions/search-chatwork/CHANGELOG.md
@@ -5,7 +5,8 @@
 - _Cache_ "Rooms" with the help of `useCachedPromise`
 - Make _OAuth_ more robust with the use of OAuth Util
 - Add **Room** icon and accessories
-<!-- - Show `EmptyView` when no rooms -->
+- Add "Search Contacts" `command` (you will need to use an API Token)
+- Remove use of `any` for better TS typing
 
 ### Dev Notes
 - Replace `for` with `map` in Rooms